+++ packages/svelte/babel/config.js
@@ -1,21 +1,22 @@
-const building = process.env.BABEL_ENV === 'build';
+const building = process.env.BABEL_ENV === "build";
 
-const plugins = [
-  'transform-object-rest-spread'
-];
+const plugins = ["transform-object-rest-spread"];
 
 if (building) {
-  plugins.push('external-helpers');
+  plugins.push("external-helpers");
 }
 
 module.exports = {
   presets: [
-    ['env', {
-      modules: building ? false : 'commonjs',
-      targets: {
-        browsers: ['> 1%']
+    [
+      "env",
+      {
+        modules: building ? false : "commonjs",
+        targets: {
+          browsers: ["> 0.25%"]
+        }
       }
-    }]
+    ]
   ],
   plugins: plugins
 };