+++ packages/svelte/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  moduleFileExtensions: ["js", "html"],
+  transform: {
+    "\\.js$": "babel-jest",
+    "\\.html$": "./scripts/svelteCompile"
+  },
+  transformIgnorePatterns: ["node_modules/(?!(svelte)/)"],
+  testMatch: ["**/tests/**/*.spec.js"]
+};