+++ packages/svelte/scripts/svelteCompile.js
@@ -1,0 +1,8 @@
+const svelte = require("svelte");
+
+exports.process = function process(src) {
+  const result = svelte.compile(src, {
+    format: "cjs"
+  });
+  return result.js.code;
+};